requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.27
requests==2.26.0
pytest==6.2.5
psycopg2-binary==2.9.3
Werkzeug==2.0.3


